smallfixes, mostly UI
