+++ public/css/main.css
@@ -285,7 +285,7 @@ TESTIMONIAL
 
 .hover:hover {
 	opacity: 0.5;
-	transition: 0.5s ease;
+	transition: 0s ease;
 	box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
 }
