+++ views/customize.hbs
@@ -23,7 +23,7 @@
 						</div>
 					</div>
 
-					<div class="server-response"></div>
+					<div class="server-response text-danger"></div>
 
 					<button class="btn btn-primary rounded-pill float-right mt-5" type="submit">
 						Save Customization
