+++ views/siteViews/home.hbs
@@ -1,15 +1,17 @@
-	<div class="container-fluid">
-		<div class="row">
-			<div class="col-md-12 splash d-flex justify-content-center align-items-center bg-primary">
-				<h1 class="p-2">Home</h1>
-			</div>
+<div class="container-fluid">
+	<div class="row">
+		<div class="col-md-12 splash d-flex justify-content-center align-items-center bg-primary">
+			<h1 class="p-2">Home</h1>
 		</div>
-
+		<div class="col-12 text-center bg-secondary p-5">
+			Secondary - This text should be visible
+		</div>
 	</div>
 
+</div>
 
 
-	<footer class="bg-light text-center pt-5 pb-5">
-		&copy; {{siteData.businessName}}, <a href="/">eEz Commerce 2020</a>
-	</footer>
 
+<footer class="bg-light text-center pt-5 pb-5">
+	&copy; {{siteData.businessName}}, <a href="/">eEz Commerce 2020</a>
+</footer>
